Refactor: svg 파일명 뒤에 -width 를 추가한다
