+++ src/components/common/Checkbox/Checkbox.component.tsx
@@ -1,6 +1,6 @@
 import React, { ChangeEventHandler } from 'react';
 import * as Styled from './Checkbox.styled';
-import CheckboxChecked from '@/assets/svg/checkbox-checked.svg';
+import CheckboxChecked from '@/assets/svg/checkbox-active-12.svg';
 
 export interface CheckboxProps extends React.LabelHTMLAttributes<HTMLLabelElement> {
   handleToggle: ChangeEventHandler<HTMLInputElement>;
